Bouton d'acceptation d'ami fonctionnel et ajout de la page Amis et gestion d'erreurs en plus

+++ backend/src/main/java/com/epic7/backend/service/MessageService.java
@@ -368,6 +368,14 @@ public class MessageService {
         if (message == null) {
             return false; // Message introuvable ou accès non autorisé
         }
+        if (message.isContainItems()) {
+            // Si le message contient des objets, on ne peut pas le supprimer
+            throw new IllegalArgumentException("Impossible de supprimer un message contenant des objets");
+        }
+        if (message.isFriendRequest()) {
+            // Si le message est une demande d'ami, on ne peut pas le supprimer
+            throw new IllegalArgumentException("Impossible de supprimer une demande d'ami");
+        }
         messageRepository.delete(message);
         return true;
     }
